validation added

+++ controllers/eventController.js
@@ -1,18 +1,18 @@
+const { validationResult } = require('express-validator');
 const { Event } = require('../models');
-const eventValidator = require('../validations/eventvalidator');
 
 const createEvent = async (req, res) => {
   try {
-    const errors = eventValidator(req.body);
-    if (errors) return res.status(404).json({ error: errors });
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) return res.status(404).json({ error: errors.array() });
     const event = await Event.create(req.body);
     return res.status(200).json({ event });
   } catch (error) {
-    return res.status(500).json({ error: 'Something Went Wrong Please Try Again' });
+    return res.status(500).json({ error });
   }
 };
 
-const getAllEvents = async (req, res) => {
+const getAllEvents = async (_req, res) => {
   try {
     const events = await Event.findAll();
     if (events.length > 0) {
@@ -27,6 +27,7 @@ const getAllEvents = async (req, res) => {
 const getEventById = async (req, res) => {
   try {
     const { id } = req.params;
+    if (!id) return res.status(404).json({ error: 'id must be number' });
     const event = await Event.findOne({ where: { id } });
     if (event) return res.status(200).json({ event });
     return res.status(404).json({ error: 'Event with this id not found' });
@@ -38,6 +39,7 @@ const getEventById = async (req, res) => {
 const updateEvent = async (req, res) => {
   try {
     const { id } = req.params;
+    if (!id) return res.status(404).json({ error: 'id must be number' });
     const [update] = await Event.update(req.body, {
       where: { id },
     });
@@ -54,6 +56,7 @@ const updateEvent = async (req, res) => {
 const deleteEvent = async (req, res) => {
   try {
     const { id } = req.params;
+    if (!id) return res.status(404).json({ error: 'id must be number' });
     const deleted = await Event.destroy({ where: { id } });
     if (deleted) return res.status(200).json({ message: 'Event Deleted Successfully' });
     return res.status(404).json({ error: 'Event with this id not found' });
